move freeport to testing
