+++ tests/callbacks_test.go
@@ -46,7 +46,7 @@ import (
 	"go.temporal.io/sdk/worker"
 	"go.temporal.io/sdk/workflow"
 	"go.temporal.io/server/common/dynamicconfig"
-	"go.temporal.io/server/common/freeport"
+	"go.temporal.io/server/common/testing/freeport"
 	"go.temporal.io/server/common/testing/protoassert"
 	"go.temporal.io/server/components/callbacks"
 	"go.temporal.io/server/tests/testcore"
